extract configudation validation from constructor
