+++ receivers/slack/config.go
@@ -1,10 +1,16 @@
 package slack
 
 import (
-	"github.com/grafana/alerting/receivers"
-)
+	"context"
+	"encoding/json"
+	"errors"
+	"fmt"
+	"net/url"
+	"strings"
 
-const FooterIconURL = "https://grafana.com/static/assets/img/fav32.png"
+	"github.com/grafana/alerting/receivers"
+	"github.com/grafana/alerting/templates"
+)
 
 type Config struct {
 	EndpointURL    string                          `json:"endpointUrl,omitempty" yaml:"endpointUrl,omitempty"`
@@ -19,4 +25,50 @@ type Config struct {
 	MentionChannel string                          `json:"mentionChannel,omitempty" yaml:"mentionChannel,omitempty"`
 	MentionUsers   receivers.CommaSeparatedStrings `json:"mentionUsers,omitempty" yaml:"mentionUsers,omitempty"`
 	MentionGroups  receivers.CommaSeparatedStrings `json:"mentionGroups,omitempty" yaml:"mentionGroups,omitempty"`
+}
+
+func ValidateConfig(factoryConfig receivers.FactoryConfig) (Config, error) {
+	decryptFunc := factoryConfig.DecryptFunc
+	var settings Config
+	err := json.Unmarshal(factoryConfig.Config.Settings, &settings)
+	if err != nil {
+		return Config{}, fmt.Errorf("failed to unmarshal settings: %w", err)
+	}
+
+	if settings.EndpointURL == "" {
+		settings.EndpointURL = APIURL
+	}
+	slackURL := decryptFunc(context.Background(), factoryConfig.Config.SecureSettings, "url", settings.URL)
+	if slackURL == "" {
+		slackURL = settings.EndpointURL
+	}
+
+	apiURL, err := url.Parse(slackURL)
+	if err != nil {
+		return Config{}, fmt.Errorf("invalid URL %q", slackURL)
+	}
+	settings.URL = apiURL.String()
+
+	settings.Recipient = strings.TrimSpace(settings.Recipient)
+	if settings.Recipient == "" && settings.URL == APIURL {
+		return Config{}, errors.New("recipient must be specified when using the Slack chat API")
+	}
+	if settings.MentionChannel != "" && settings.MentionChannel != "here" && settings.MentionChannel != "channel" {
+		return Config{}, fmt.Errorf("invalid value for mentionChannel: %q", settings.MentionChannel)
+	}
+	settings.Token = decryptFunc(context.Background(), factoryConfig.Config.SecureSettings, "token", settings.Token)
+	if settings.Token == "" && settings.URL == APIURL {
+		return Config{}, errors.New("token must be specified when using the Slack chat API")
+	}
+	if settings.Username == "" {
+		settings.Username = "Grafana"
+	}
+	if settings.Text == "" {
+		settings.Text = templates.DefaultMessageEmbed
+	}
+	if settings.Title == "" {
+		settings.Title = templates.DefaultMessageTitleEmbed
+	}
+
+	return settings, nil
 }
